Mixin no longer a mixin
